apps/Versionamento/requirements.txt
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.32.3
urllib3==2.2.3
pandas==2.2.2
geopandas==0.14.4
pyarrow==16.1.0
azure-storage-blob==12.19.1
